requirements.txt
alpaca-py
pandas
numpy
schedule
lxml
html5lib
beautifulsoup4
pytz
yfinance